+++ golang/lib/client/metrics_client_creator.go
@@ -5,6 +5,7 @@ import (
 	"github.com/kurtosis-tech/metrics-library/golang/lib/client/segment_client"
 	"github.com/kurtosis-tech/metrics-library/golang/lib/source"
 	"github.com/kurtosis-tech/stacktrace"
+	"gopkg.in/segmentio/analytics-go.v3"
 )
 
 const(
@@ -13,7 +14,7 @@ const(
 
 //The argument shouldFlushQueueOnEachEvent is used to imitate a sync request, it is not exactly the same because
 //the event is enqueued but the queue is flushed suddenly so is pretty close to event traked in sync
-func CreateMetricsClient(source source.Source, sourceVersion string, userId string, didUserAcceptSendingMetrics bool, shouldFlushQueueOnEachEvent bool) (MetricsClient, error) {
+func CreateMetricsClient(source source.Source, sourceVersion string, userId string, didUserAcceptSendingMetrics bool, shouldFlushQueueOnEachEvent bool, callback analytics.Callback) (MetricsClient, error) {
 
 	metricsClientType := DoNothing
 
@@ -23,7 +24,7 @@ func CreateMetricsClient(source source.Source, sourceVersion string, userId stri
 
 	switch metricsClientType {
 	case Segment:
-		metricsClient, err := segment_client.NewSegmentClient(source, sourceVersion, userId, shouldFlushQueueOnEachEvent)
+		metricsClient, err := segment_client.NewSegmentClient(source, sourceVersion, userId, shouldFlushQueueOnEachEvent, callback)
 		if err != nil {
 			return nil, stacktrace.Propagate(err, "An error occurred creating Segment metrics client")
 		}